gs: Harmonize mock server usage in tests
