+++ pkg/gatewayserver/io/basicstationlns/basicstationlns_util_test.go
@@ -16,8 +16,10 @@ package basicstationlns_test
 
 import (
 	"context"
+	"time"
 
 	"go.thethings.network/lorawan-stack/pkg/auth/rights"
+	"go.thethings.network/lorawan-stack/pkg/component"
 	"go.thethings.network/lorawan-stack/pkg/rpcmetadata"
 	"go.thethings.network/lorawan-stack/pkg/ttnpb"
 	"go.thethings.network/lorawan-stack/pkg/unique"
@@ -39,4 +41,14 @@ func newContextWithRightsFetcher(ctx context.Context) context.Context {
 			return
 		}),
 	)
+}
+
+func mustHavePeer(ctx context.Context, c *component.Component, role ttnpb.PeerInfo_Role) {
+	for i := 0; i < 20; i++ {
+		time.Sleep(20 * time.Millisecond)
+		if peer := c.GetPeer(ctx, role, nil); peer != nil {
+			return
+		}
+	}
+	panic("could not connect to peer")
 }
