+++ pkg/gatewayserver/io/mqtt/mqtt_util_test.go
@@ -16,16 +16,10 @@ package mqtt_test
 
 import (
 	"context"
-	"fmt"
-	"net"
 	"time"
 
 	"go.thethings.network/lorawan-stack/pkg/component"
-	"go.thethings.network/lorawan-stack/pkg/errors"
-	"go.thethings.network/lorawan-stack/pkg/rpcserver"
 	"go.thethings.network/lorawan-stack/pkg/ttnpb"
-	"go.thethings.network/lorawan-stack/pkg/unique"
-	"google.golang.org/grpc/metadata"
 )
 
 func mustHavePeer(ctx context.Context, c *component.Component, role ttnpb.PeerInfo_Role) {
@@ -36,72 +30,4 @@ func mustHavePeer(ctx context.Context, c *component.Component, role ttnpb.PeerIn
 		}
 	}
 	panic("could not connect to peer")
-}
-
-type mockIS struct {
-	ttnpb.GatewayRegistryServer
-	ttnpb.GatewayAccessServer
-	gateways     map[string]*ttnpb.Gateway
-	gatewayAuths map[string][]string
-}
-
-func startMockIS(ctx context.Context) (*mockIS, string) {
-	is := &mockIS{
-		gateways:     make(map[string]*ttnpb.Gateway),
-		gatewayAuths: make(map[string][]string),
-	}
-	srv := rpcserver.New(ctx)
-	ttnpb.RegisterGatewayRegistryServer(srv.Server, is)
-	ttnpb.RegisterGatewayAccessServer(srv.Server, is)
-	lis, err := net.Listen("tcp", ":0")
-	if err != nil {
-		panic(err)
-	}
-	go srv.Serve(lis)
-	return is, lis.Addr().String()
-}
-
-func (is *mockIS) add(ctx context.Context, ids ttnpb.GatewayIdentifiers, key string) {
-	uid := unique.ID(ctx, ids)
-	is.gateways[uid] = &ttnpb.Gateway{
-		GatewayIdentifiers: ids,
-	}
-	if key != "" {
-		is.gatewayAuths[uid] = []string{fmt.Sprintf("Bearer %v", key)}
-	}
-}
-
-var errNotFound = errors.DefineNotFound("not_found", "not found")
-
-func (is *mockIS) Get(ctx context.Context, req *ttnpb.GetGatewayRequest) (*ttnpb.Gateway, error) {
-	uid := unique.ID(ctx, req.GatewayIdentifiers)
-	app, ok := is.gateways[uid]
-	if !ok {
-		return nil, errNotFound
-	}
-	return app, nil
-}
-
-func (is *mockIS) ListRights(ctx context.Context, ids *ttnpb.GatewayIdentifiers) (res *ttnpb.Rights, err error) {
-	res = &ttnpb.Rights{}
-	md, ok := metadata.FromIncomingContext(ctx)
-	if !ok {
-		return
-	}
-	authorization, ok := md["authorization"]
-	if !ok || len(authorization) == 0 {
-		return
-	}
-	auths, ok := is.gatewayAuths[unique.ID(ctx, *ids)]
-	if !ok {
-		return
-	}
-	for _, auth := range auths {
-		if auth == authorization[0] {
-			res.Rights = append(res.Rights,
-				ttnpb.RIGHT_GATEWAY_LINK,
-			)
-		}
-	}
-	return
 }
